Move otlp utilities to otlp package
