+++ exporters/otlp/all/src/main/java/io/opentelemetry/exporter/otlp/internal/GrpcExporter.java
@@ -3,60 +3,57 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.exporter.internal.http;
+package io.opentelemetry.exporter.otlp.internal;
 
 import io.opentelemetry.api.metrics.MeterProvider;
 import io.opentelemetry.exporter.internal.FailedExportException;
-import io.opentelemetry.exporter.internal.grpc.GrpcExporterUtil;
 import io.opentelemetry.exporter.internal.marshal.Marshaler;
 import io.opentelemetry.exporter.internal.metrics.ExporterInstrumentation;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.common.InternalTelemetryVersion;
-import io.opentelemetry.sdk.common.export.HttpResponse;
-import io.opentelemetry.sdk.common.export.HttpSender;
-import io.opentelemetry.sdk.common.export.MessageWriter;
+import io.opentelemetry.sdk.common.export.GrpcResponse;
+import io.opentelemetry.sdk.common.export.GrpcSender;
+import io.opentelemetry.sdk.common.export.GrpcStatusCode;
 import io.opentelemetry.sdk.common.internal.StandardComponentId;
 import io.opentelemetry.sdk.common.internal.ThrottlingLogger;
-import java.io.IOException;
 import java.net.URI;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Supplier;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import javax.annotation.Nullable;
 
 /**
- * An exporter for http/protobuf or http/json using a signal-specific Marshaler.
+ * Generic gRPC exporter.
  *
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
  */
 @SuppressWarnings("checkstyle:JavadocMethod")
-public final class HttpExporter {
+public final class GrpcExporter {
 
-  private static final Logger internalLogger = Logger.getLogger(HttpExporter.class.getName());
+  private static final Logger internalLogger = Logger.getLogger(GrpcExporter.class.getName());
 
   private final ThrottlingLogger logger = new ThrottlingLogger(internalLogger);
+
+  // We only log unimplemented once since it's a configuration issue that won't be recovered.
+  private final AtomicBoolean loggedUnimplemented = new AtomicBoolean();
   private final AtomicBoolean isShutdown = new AtomicBoolean();
 
   private final String type;
-  private final HttpSender httpSender;
+  private final GrpcSender grpcSender;
   private final ExporterInstrumentation exporterMetrics;
-  private final boolean exportAsJson;
 
-  public HttpExporter(
-      StandardComponentId componentId,
-      HttpSender httpSender,
-      Supplier<MeterProvider> meterProviderSupplier,
+  public GrpcExporter(
+      GrpcSender grpcSender,
       InternalTelemetryVersion internalTelemetryVersion,
-      URI endpoint,
-      boolean exportAsJson) {
+      StandardComponentId componentId,
+      Supplier<MeterProvider> meterProviderSupplier,
+      URI endpoint) {
     this.type = componentId.getStandardType().signal().logFriendlyName();
-    this.httpSender = httpSender;
+    this.grpcSender = grpcSender;
     this.exporterMetrics =
         new ExporterInstrumentation(
             internalTelemetryVersion, meterProviderSupplier, componentId, endpoint);
-    this.exportAsJson = exportAsJson;
   }
 
   public CompletableResultCode export(Marshaler exportRequest, int numItems) {
@@ -68,12 +65,10 @@ public final class HttpExporter {
         exporterMetrics.startRecordingExport(numItems);
 
     CompletableResultCode result = new CompletableResultCode();
-    MessageWriter messageWriter =
-        exportAsJson ? exportRequest.toJsonMessageWriter() : exportRequest.toBinaryMessageWriter();
 
-    httpSender.send(
-        messageWriter,
-        httpResponse -> onResponse(result, metricRecording, httpResponse),
+    grpcSender.send(
+        exportRequest.toBinaryMessageWriter(),
+        grpcResponse -> onResponse(result, metricRecording, grpcResponse),
         throwable -> onError(result, metricRecording, throwable));
 
     return result;
@@ -82,33 +77,47 @@ public final class HttpExporter {
   private void onResponse(
       CompletableResultCode result,
       ExporterInstrumentation.Recording metricRecording,
-      HttpResponse httpResponse) {
-    int statusCode = httpResponse.getStatusCode();
+      GrpcResponse grpcResponse) {
+    GrpcStatusCode statusCode = grpcResponse.getStatusCode();
 
-    metricRecording.setHttpStatusCode(statusCode);
+    metricRecording.setGrpcStatusCode(statusCode);
 
-    if (statusCode >= 200 && statusCode < 300) {
+    if (statusCode == GrpcStatusCode.OK) {
       metricRecording.finishSuccessful();
       result.succeed();
       return;
     }
 
-    metricRecording.finishFailed(String.valueOf(statusCode));
-
-    byte[] body = httpResponse.getResponseBody();
-
-    String status = extractErrorStatus(httpResponse.getStatusMessage(), body);
-
-    logger.log(
-        Level.WARNING,
-        "Failed to export "
-            + type
-            + "s. Server responded with HTTP status code "
-            + statusCode
-            + ". Error message: "
-            + status);
-
-    result.failExceptionally(FailedExportException.httpFailedWithResponse(httpResponse));
+    metricRecording.finishFailed(String.valueOf(statusCode.getValue()));
+    switch (statusCode) {
+      case UNIMPLEMENTED:
+        if (loggedUnimplemented.compareAndSet(false, true)) {
+          GrpcExporterUtil.logUnimplemented(
+              internalLogger, type, grpcResponse.getStatusDescription());
+        }
+        break;
+      case UNAVAILABLE:
+        logger.log(
+            Level.SEVERE,
+            "Failed to export "
+                + type
+                + "s. Server is UNAVAILABLE. "
+                + "Make sure your collector is running and reachable from this network. "
+                + "Full error message:"
+                + grpcResponse.getStatusDescription());
+        break;
+      default:
+        logger.log(
+            Level.WARNING,
+            "Failed to export "
+                + type
+                + "s. Server responded with gRPC status code "
+                + statusCode.getValue()
+                + ". Error message: "
+                + grpcResponse.getStatusDescription());
+        break;
+    }
+    result.failExceptionally(FailedExportException.grpcFailedWithResponse(grpcResponse));
   }
 
   private void onError(
@@ -120,10 +129,13 @@ public final class HttpExporter {
         Level.SEVERE,
         "Failed to export "
             + type
-            + "s. The request could not be executed. Full error message: "
+            + "s. The request could not be executed. Error message: "
             + e.getMessage(),
         e);
-    result.failExceptionally(FailedExportException.httpFailedExceptionally(e));
+    if (logger.isLoggable(Level.FINEST)) {
+      logger.log(Level.FINEST, "Failed to export " + type + "s. Details follow:", e);
+    }
+    result.failExceptionally(FailedExportException.grpcFailedExceptionally(e));
   }
 
   public CompletableResultCode shutdown() {
@@ -131,17 +143,6 @@ public final class HttpExporter {
       logger.log(Level.INFO, "Calling shutdown() multiple times.");
       return CompletableResultCode.ofSuccess();
     }
-    return httpSender.shutdown();
-  }
-
-  private static String extractErrorStatus(String statusMessage, @Nullable byte[] responseBody) {
-    if (responseBody == null) {
-      return "Response body missing, HTTP status message: " + statusMessage;
-    }
-    try {
-      return GrpcExporterUtil.getStatusMessage(responseBody);
-    } catch (IOException e) {
-      return "Unable to parse response body, HTTP status message: " + statusMessage;
-    }
+    return grpcSender.shutdown();
   }
 }
