+++ exporters/otlp/all/src/main/java/io/opentelemetry/exporter/otlp/internal/GzipCompressor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.exporter.internal.compression;
+package io.opentelemetry.exporter.otlp.internal;
 
 import io.opentelemetry.sdk.common.export.Compressor;
 import java.io.IOException;
