+++ exporters/otlp/all/src/test/java/io/opentelemetry/exporter/otlp/internal/GrpcExporterTest.java
@@ -3,11 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.exporter.internal.grpc;
+package io.opentelemetry.exporter.otlp.internal;
 
 import static io.opentelemetry.sdk.common.export.GrpcStatusCode.UNAVAILABLE;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doAnswer;
 
@@ -26,31 +25,13 @@ import io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions;
 import io.opentelemetry.sdk.testing.exporter.InMemoryMetricReader;
 import java.io.IOException;
 import java.net.URI;
-import java.time.Duration;
 import java.util.function.Consumer;
 import javax.annotation.Nullable;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.Mockito;
 
 class GrpcExporterTest {
-
-  @Test
-  void build_NoGrpcSenderProvider() {
-    assertThatThrownBy(
-            () ->
-                new GrpcExporterBuilder(
-                        StandardComponentId.ExporterType.OTLP_GRPC_SPAN_EXPORTER,
-                        Duration.ofSeconds(10),
-                        new URI("http://localhost"),
-                        "service/method")
-                    .build())
-        .isInstanceOf(IllegalStateException.class)
-        .hasMessage(
-            "No GrpcSenderProvider found on classpath. Please add dependency on "
-                + "opentelemetry-exporter-sender-okhttp or opentelemetry-exporter-sender-grpc-managed-channel");
-  }
 
   @ParameterizedTest
   @EnumSource
