+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/OtlpGrpcProfileExporter.java
@@ -5,8 +5,8 @@
 
 package io.opentelemetry.exporter.otlp.profiles;
 
-import io.opentelemetry.exporter.internal.grpc.GrpcExporter;
-import io.opentelemetry.exporter.internal.grpc.GrpcExporterBuilder;
+import io.opentelemetry.exporter.otlp.internal.GrpcExporter;
+import io.opentelemetry.exporter.otlp.internal.GrpcExporterBuilder;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import java.util.Collection;
 import java.util.StringJoiner;
